Move cross-feature phrase joins to phrases/live.ts using nested q.from()
Validates @tanstack/db v0.6's "includes" pattern (nested q.from() inside
.select()) on usePhrasePlaylists and usePhraseComments. The outer query
yields the phrase row; the playlists/comments fields are child query
builders that the type system maps to Collection<...>. The hook internally
subscribes to the child collection with useLiveQuery to preserve its flat
external API (no consumer changes required).

Side effects:
- phrases/hooks.ts no longer imports other features' collections.ts.
  The cross-feature imports now live in phrases/live.ts, which is the
  named home for this kind of join per #593's architectural rule.
- useRelatedCards moves to live.ts too — its sibling-phrase queries
  still need playlistPhraseLinksCollection / phraseRequestsCollection.
- big-phrase-card imports from the @/features/phrases barrel.

usePhraseProvenance is unchanged in shape; it still merges the two flat
arrays in JS for the existing UI. A future consumer-level refactor could
pass child collections down into per-section components for fine-grained
reactivity, but that's out of scope for this spike.

Refs #593.
