+++ src/components/cards/big-phrase-card.tsx
@@ -43,7 +43,7 @@ import {
 	usePhraseProvenance,
 	useRelatedCards,
 	type PhraseProvenanceItem as PhraseProvenanceItemType,
-} from '@/features/phrases/hooks'
+} from '@/features/phrases'
 import { PhraseTinyCard } from '@/components/cards/phrase-tiny-card'
 import { PlaylistEmbed } from '@/components/playlists/playlist-embed'
 import Flagged from '@/components/flagged'
